Solved bug #60 AVRDude verbose option breaks programming
The verbose option of avrdude was outputting a line like this
  avrdude: usbdev_open(): Found AVRISP mkII, serno: 000200129194
and that was seen as a fatal error because the word usbdev_open
was found.
The change has also been made in AvariceOutputListener.java but it
has neither been tested nor do I know if the bug exists for avarice.