Add a few more default toolchain options (-ffunction-sections, etc)
To make it more convenient to start a new project some options are
now switched on by default. The options are mostly needed to create
a size effecient binary.
The following options can be set in the Compiler Optimization settings
  -ffunction-sections
  -fdata-sections
The linker has --gc-sections on by default, so that the two compiler
options above take their effect.
In the linker section -lm now also has a checkbox (default on).