Remove the linker option for -lm
It could have been a good idea, but with MBS there is no simple method
to add the -lm option at the end of the command. The LIBS and other OBJS
are treated special internally, and there is nothing you can specify
in plugin.xml to do what we want.

Perhaps a dedicated valueHandler, but right now I have no clue if that
can will be enough.