Handle avrdude.conf from avrdude 6.1 too
With the new version of avrdude the syntax of avrdude.conf was changed.
Both "programmer" and "part" can have a parent. To handle that new
structure all lines of avrdude.conf are read and the programmer config
entries are assembled from that.
This commit only changed the parsing of programmer information. "Part"
will be done later.